Move lock from 'discrete_domaines' to 'utils'.
